corrected imports for Navigraph API
